fix(store): moved store.ts and slices directory to a separate folder
